+++ api/src/main/java/grooteogi/service/PostService.java
@@ -14,7 +14,7 @@ public class PostService {
 
   private final PostRepository postRepository;
 
-  public CursorResult<Post> search(String search, Long cursorId, String type,
+  public CursorResult<Post> search(String search, Integer cursorId, String type,
       Pageable page) {
     final List<Post> posts;
     if (search == null) {
@@ -22,22 +22,22 @@ public class PostService {
     } else {
       posts = getPosts(search, cursorId, page, type);
     }
-    final Long lastIdOfList = posts.isEmpty() ? null : posts.get(posts.size() - 1).getId();
+    final Integer lastIdOfList = posts.isEmpty() ? null : posts.get(posts.size() - 1).getId();
     return new CursorResult<>(posts, hasNext(lastIdOfList));
   }
 
-  private List<Post> getAllPosts(Long cursorId, Pageable page, String type) {
+  private List<Post> getAllPosts(Integer cursorId, Pageable page, String type) {
     return cursorId == 0 ? this.postRepository.findAllByOrderByIdDesc(page) :
         this.postRepository.findByIdLessThanOrderByIdDesc(cursorId, page);
   }
 
-  private List<Post> getPosts(String search, Long cursorId, Pageable page, String type) {
+  private List<Post> getPosts(String search, Integer cursorId, Pageable page, String type) {
     return cursorId == 0 ? this.postRepository
         .findByTitleContainingOrContentContainingOrderByIdDesc(search, search, page) :
         this.postRepository.findBySearchOrderByIdDesc(search, search, cursorId, page);
   }
 
-  private Boolean hasNext(Long lastIdOfList) {
+  private Boolean hasNext(Integer lastIdOfList) {
     if (lastIdOfList == null) {
       return false;
     }
